+++ test/flake.nix
@@ -24,21 +24,54 @@
 
         systems = [ "x86_64-linux" ];
       }
-      ({ make-shell, pkgs, ps-tools, purs-nix, ... }:
+      ({ make-shell, pkgs, ps-tools, system, ... }:
          let
+           purs-nix =
+             purs-nix-flake
+               { inherit system;
+
+                 overlays =
+                   [ (_: super: assert !super?test-package; {})
+
+                     (self: super:
+                        with self;
+                        { inherit (super) console;
+
+                          effect =
+                            { src.path = purs-nix-test-packages;
+                              info = /effect/package.nix;
+                            };
+
+                          # testing overlay precedence
+                          murmur3 = prelude;
+
+                          # test if later overlays have access to this package
+                          test-package = prelude;
+                        }
+                     )
+
+                     (self: super:
+                        with self;
+                        { murmur3 = self."ursi.murmur3";
+                          "test.prelude" = super.test-package;
+                        }
+                     )
+                   ];
+               };
+
            minimal = false;
 
            switches =
              mapAttrs
                (n: v: (a: l.warnIf (!a) "switches.${n} == false" a) (!minimal && v))
                { packages-compile = true;
-                 parser = true;
+                 parser = false;
                  repl = true;
                };
 
            l = p.lib; p = pkgs;
            inherit (purs-nix) ps-pkgs ps-pkgs-ns purs;
-           package = import ./package.nix purs-nix-test-packages purs-nix;
+           package = import ./package.nix purs-nix;
 
            ps-custom = { dir ? ./., ...}@args:
              purs
@@ -67,12 +100,12 @@
            ps2 =
              purs
                { dependencies =
-                   let inherit (ps-pkgs-ns) ursi; in
+                   let inherit (ps-pkgs-ns) test; in
                    with ps-pkgs;
                    [ console
                      effect
-                     prelude
-                     ursi.murmur3
+                     murmur3
+                     test.prelude
                    ];
 
                   test-dependencies = [ ps-pkgs."assert" ];
@@ -97,7 +130,7 @@
                ({ inherit l p switches;
                   inherit (purs-nix-flake.inputs.parsec.lib) parsec;
                 }
-                // purs-nix
+                // (purs-nix-flake { inherit system; })
                );
 
            expected-output-tail =